+++ Dockerfile
@@ -2,4 +2,4 @@ FROM mcr.microsoft.com/azure-functions/node:4-node16-slim
 
 ENV AzureWebJobsScriptRoot=/home/site/wwwroot
 
-COPY . .
+COPY ./src .
