--- color/sample.dat
@@ -1,3 +1,0 @@
-{
-    "name": "Azure"
-}
