--- colors/sample.dat
@@ -1,5 +1,0 @@
-{
-    "id": "123",
-    "lat":1,
-    "long":2 
-}
