--- local.settings.json
@@ -1,6 +1,0 @@
-{
-  "IsEncrypted": false,
-  "Values": {
-    "FUNCTIONS_WORKER_RUNTIME": "node"
-  }
-}
