--- package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "applicationinsights": "^2.3.1",
-    "azure-functions-core-tools": "^3.0.4425",
-    "redis": "^4.0.6"
-  }
-}
