+++ src/package.json
@@ -1,0 +1,17 @@
+{
+
+  "name": "viridis",
+  "version": "alpha-0.0.1",
+  "description": "viridis, colors",
+  "main": "index.js",
+  "scripts": {
+    "start": "node index.js"
+  },
+  "author": "",
+  "dependencies": {
+    "express":"^4.17.3",
+    "applicationinsights": "^2.3.1",
+    "redis": "^4.0.6",
+    "fs": "^0.0.1-security"
+  }
+}
